feat: add FeedbackImage component and integrate it into ImagePreviewButton and SheetDetailTable
